remove config + output files with hardcoded db ids
Removes config and output files such as
"variable_replacemennts.json" and
"charts_to_update.json" that contain hard-coded SQL
db ids. The problem with hard-coding these db ids is
that the files may havve been constructed from a SQL
db instance that was not up to date with the
production db.